# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SSC GD Smart Coach API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions with filters
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: topic
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: isPYQ
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedQuestions"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question with solution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
  /questions/{id}/bookmark:
    post:
      operationId: bookmarkQuestion
      tags: [questions]
      summary: Bookmark a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bookmark status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkStatus"
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List study notes/PDFs
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: chapter
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: isPremium
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get note details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Note detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}/bookmark:
    post:
      operationId: bookmarkNote
      tags: [notes]
      summary: Bookmark a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bookmark status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkStatus"
...
  /quiz/daily/submit:
    post:
      operationId: submitDailyQuiz
      tags: [quiz]
      summary: Submit daily quiz answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /quiz/current-affairs:
    get:
      operationId: getCurrentAffairs
      tags: [quiz]
      summary: Get current affairs questions
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Current affairs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"

  # ─── VIDEOS ──────────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkStatus"

  # ─── ADMIN ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Question:
      type: object
      required: [id, text, options, correctOption, subject, difficulty]
      properties:
        id:             { type: string }
        text:           { type: string }
        options:
          type: array
          items:
            type: object
            required: [id, text]
            properties:
              id:   { type: string }
              text: { type: string }
        correctOption:  { type: string }
        explanation:    { type: ["string", "null"] }
        subject:        { type: string }
        topic:          { type: ["string", "null"] }
        difficulty:     { type: string, enum: [easy, medium, hard] }
        year:           { type: ["integer", "null"] }
        isPYQ:          { type: boolean }
        isBookmarked:   { type: boolean }
    QuestionDetail:
      type: object
      required: [id, text, options, correctOption, explanation, subject, difficulty]
      properties:
        id:            { type: string }
        text:          { type: string }
        options:
          type: array
          items:
            type: object
            required: [id, text]
            properties:
              id:   { type: string }
              text: { type: string }
        correctOption: { type: string }
        explanation:   { type: string }
        subject:       { type: string }
        topic:         { type: ["string", "null"] }
        difficulty:    { type: string }
        year:          { type: ["integer", "null"] }
        isPYQ:         { type: boolean }
        isBookmarked:  { type: boolean }
        aiExplanation: { type: ["string", "null"] }
    PaginatedQuestions:
      type: object
      required: [questions, total, page, limit]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
...
    BookmarkStatus:
      type: object
      required: [success, isBookmarked]
      properties:
        success:    { type: boolean }
        isBookmarked: { type: boolean }

    # ─── Note ───────────────────────────────────────────────────────────────
    Note:
      type: object
      required: [id, title, subject, chapter, fileUrl, isPremium, downloadCount]
      properties:
        id:            { type: string }
        title:         { type: string }
        subject:       { type: string }
        chapter:       { type: string }
        description:   { type: ["string", "null"] }
        fileUrl:       { type: string }
        thumbnailUrl:  { type: ["string", "null"] }
        fileSize:      { type: ["string", "null"] }
        isPremium:     { type: boolean }
        downloadCount: { type: integer }
        isBookmarked:  { type: boolean }
        createdAt:     { type: string }
...
    QuizSubmission:
      type: object
      required: [quizId, answers]
      properties:
        quizId: { type: string }
        answers:
          type: array
          items:
            type: object
            required: [questionId, selectedOption]
            properties:
              questionId:     { type: string }
              selectedOption: { type: ["string", "null"] }
    QuizResult:
      type: object
      required: [quizId, score, total, correct, incorrect, streakUpdated]
      properties:
        quizId:        { type: string }
        score:         { type: number }
        total:         { type: integer }
        correct:       { type: integer }
        incorrect:     { type: integer }
        streakUpdated: { type: boolean }
        newStreak:     { type: integer }
        answers:
          type: array
          items:
            type: object
            required: [questionId, isCorrect, correctOption, explanation]
            properties:
              questionId:     { type: string }
              isCorrect:      { type: boolean }
              correctOption:  { type: string }
              explanation:    { type: string }

    # ─── Video ───────────────────────────────────────────────────────────────
...